+++ src/BankAccountNormalizedInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Cs278\BankModulus;
+
+interface BankAccountNormalizedInterface extends BankAccountInterface
+{
+    /** @return BankAccountInterface */
+    public function getOriginalBankAccount();
+}